% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountPersonTime.R
\name{CountPersonTime}
\alias{CountPersonTime}
\title{import data.table}
\usage{
CountPersonTime(
  Dataset_events,
  Dataset,
  Person_id,
  Start_study_time,
  End_study_time,
  Start_date,
  End_date,
  Birth_date = NULL,
  Strata = NULL,
  Outcomes,
  Name_event,
  Date_event,
  Age_bands = NULL,
  Unit_of_age = "year",
  Increment = "year",
  include_remaning_ages = T,
  Aggregate = T
)
}
\description{
import data.table
}
